refactor: split pages code (#215)
* Split use mint

* Fix error to refresh cm and show unminted nfts

* Add error stated in update and create cm

* Fix error cm

* Remove network variable because is unused

* Rename previewNft and mintNft

* Remove React in import

* Remove link in nft collection

* check idl

* reset nft collection to refresh nfts

* feat: make sure NFTs are shown and improve code

* fix: remove unnecessary import
