+++ hooks/index.ts
@@ -7,3 +7,4 @@ export { default as useUploadFiles } from './useUploadFiles'
 export { default as useVerifyCandyMachineV2 } from './useVerifyCandyMachineV2'
 export { default as useMintCandyMachine } from './useMintCandyMachine'
 export { default as useNotification } from './useNotification'
+export { default as useRefreshCandyMachine } from './useRefreshCandyMachine'
